requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Limiter==3.3.1
PyJWT==2.7.0
pymongo==4.3.3
gunicorn==20.1.0
dnspython==2.2.1
Flask-Mail==0.9.1